+++ src/combinators/merge.js
@@ -4,24 +4,9 @@ import Signal from '../Signal'
  * Merges the signals `ss` and emits their values. The returned signal will
  * complete once *all* of the given signals have completed.
  *
- * @param {Array} ss The signals to merge.
- * @returns {Signal} A new signal.
- * @example
- *
- * import { Signal, merge } from 'bulb'
- *
- * const s = Signal.fromArray([1, 2, 3])
- * const t = Signal.fromArray([4, 5, 6])
- * const u = merge(s, t)
- *
- * u.subscribe(console.log) // 1, 4, 2, 5, 3, 6
+ * @private
  */
-export default function merge (...ss) {
-  // Allow the signals to be given as an array.
-  if (ss.length === 1 && Array.isArray(ss[0])) {
-    ss = ss[0]
-  }
-
+export default function merge (ss) {
   return new Signal(emit => {
     let n = 0